+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/client/TigerTracingDto.java
@@ -13,7 +13,6 @@ import lombok.Data;
 @Builder
 public class TigerTracingDto {
 
-    private final String uuid;
     private final String requestUuid;
     private final String responseUuid;
     private final RbelHostname sender;